finished midterm projects

+++ project/project_nebula/project-install.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-PROJNAME="project_nebula"
-PROJDIR="/home/masterdesky/GitHub/ELTE_Modelling_Lab_2021/project/project_nebula"
-GEANT4INSTALL="/home/masterdesky/apps/geant4/install"
+PROJNAME="ex_1"
+PROJDIR="/home/masterdesky/GitHub/ELTE_Modelling_Lab_2021/project/ex_1"
+GEANT4INSTALL="/home/masterdesky/apps/Geant4/install"
 
 # Source Geant4
 if [ -f ${GEANT4INSTALL}/bin/geant4.sh ];
@@ -12,8 +12,13 @@ then
     cd ~-
 fi
 
+if [ ! -d ${PROJDIR}/NEBULA_build ];
+then
+  mkdir -p ${PROJDIR}/NEBULA_build
+else
+  rm -rf ${PROJDIR}/NEBULA_build/*
+fi
 cd ${PROJDIR}/NEBULA_build
-rm -rf ${PROJDIR}/NEBULA_build/*
 cmake -DGeant4_DIR=${GEANT4INSTALL}/lib/Geant4-10.7.1/ \
 			${PROJDIR}/NEBULA \
 			|& tee >(ts "[%x %X]" > ${PROJDIR}/NEBULA_build/c.log)
